backend/requirements.txt
fastapi
uvicorn
google-genai
pydantic
requests
python-dotenv
sse-starlette
beautifulsoup4
duckduckgo-search
